refactor(engineer_plan): promote to directory, split tests (#1266)
- src/engineer_plan.rs: 699 → src/engineer_plan/mod.rs (311 LOC)
- New src/engineer_plan/tests.rs (388 LOC)
- Updated include_str! path (../prompt_assets → ../../prompt_assets)

Pure refactor: 37 tests pass.
